Trait algo::ecdsa::EcdsaVerify

pub trait EcdsaVerify: ErrorType {
    type PublicKey;
    type Curve: EcdsaCurve;
    type Signature;

    // Required method
    fn verify<H: HashMarker>(
        curve: &Self::Curve,
        public_key: &Self::PublicKey,
        message_hash: impl AsRef<[u8]>,
        signature: &Self::Signature,
    ) -> Result<(), Self::Error>;
}

Trait for ECDSA verification.

This trait defines the methods required for verifying ECDSA signatures.

Required Associated Types

type PublicKey

type Curve: EcdsaCurve

type Signature

Required Methods

fn verify<H: HashMarker>( curve: &Self::Curve, public_key: &Self::PublicKey, message_hash: impl AsRef<[u8]>, signature: &Self::Signature, ) -> Result<(), Self::Error>

Verifies an ECDSA signature.

Parameters

curve: The elliptic curve to use for verification.

public_key: The public key to use for verification.

message_hash: The hash of the message to verify.

signature: The signature to verify.

Returns

A result indicating whether the signature is valid, or an error.

Object Safety

This trait is not object safe.