Trait algo::mac::Mac

pub trait Mac: ErrorType {
    type InitParams;

    // Required methods
    fn init(init_params: Self::InitParams) -> Result<(), Self::Error>;
    fn set_key(&mut self, key: &[u8]) -> Result<(), Self::Error>;
    fn update(&mut self, input: &mut [u8]) -> Result<(), Self::Error>;
    fn reset(&mut self) -> Result<(), Self::Error>;
    fn finalize(&mut self, out: &mut [u8]) -> Result<(), Self::Error>;
    fn verify(&mut self, tag: &[u8]) -> Result<(), Self::Error>;
}

Required Associated Types

type InitParams

Required Methods

fn init(init_params: Self::InitParams) -> Result<(), Self::Error>

Init instance of the crypto function with the given context.

Parameters

init_params: The context or configuration parameters for the crypto function.

Returns

A new instance of the hash function.

fn set_key(&mut self, key: &[u8]) -> Result<(), Self::Error>

Sets the key for the HMAC algorithm.

Parameters

key: The key to be used for HMAC.

Returns

A Result indicating success or failure. On success, returns Ok(()). On failure, returns an error of type Self::Error.

fn update(&mut self, input: &mut [u8]) -> Result<(), Self::Error>

Update state using provided input data.

Parameters

input: The input data to be hashed. This can be any type that implements AsRef<[u8]>.

Returns

A Result indicating success or failure. On success, returns Ok(()). On failure, returns a CryptoError.

fn reset(&mut self) -> Result<(), Self::Error>

Reset instance to its initial state.

Returns

A Result indicating success or failure. On success, returns Ok(()). On failure, returns a CryptoError.

fn finalize(&mut self, out: &mut [u8]) -> Result<(), Self::Error>

Finalize the computation and produce the output.

Parameters

out: A mutable slice to store the hash output. The length of the slice must be at least MAX_OUTPUT_SIZE.

Returns

A Result indicating success or failure. On success, returns Ok(()). On failure, returns a CryptoError.

fn verify(&mut self, tag: &[u8]) -> Result<(), Self::Error>

Verifies if the given MAC tag matches the expected result.

Parameters

tag: The MAC tag to be verified.

Returns

A Result indicating success or failure. On success, returns Ok(()). On failure, returns an error of type Self::Error.

Object Safety

This trait is not object safe.